Updated default namespace/ Made the behaviour system modular
The default namespace is updated to comply with the Microsoft standard C# naming conventions.

The Behaviour System is updated to use more than one Behaviour Manager. This way, it creates less or easier to deal with merge conflicts, because all the new logic is put away in a new Behaviour Manager and not in de GameManager itself. You just have to add the new Behaviour Manager to the list in the GameManager.

I also added the DebugBehaviour and DebugBehaviourManager classes. These classes serve as an example of how to use the new Behaviour System and to debug when the game has started and ended.