Updated Inventory System/ Updated the LostAndFound/ Fixed scriptable object bug
The Inventory System is updated to now add/remove items to the inventory and also create items using items from the inventory. The Inventory System also has a UI that gets updated every time something gets added or removed from the inventory.

The scratchpad, called the LostAndFound, is updated to contain a generic version of the Retrieve method to more easily retrieve the desired object as the prefered type.

The scriptable objects (Behaviour Data) are now instantiated at the beginning of the game. Previously this was not done and therefore the Unity asset itself was affected and not a copy of it.